refactor: cleanup resolution by pulling them into own file
The build initializer now uses resolves from `./future-resolvers.ts`.
