+++ packages/core/src/new-api/internal/new-execution/future-processor/helpers/build-initialization-message-for.ts
@@ -1,27 +1,18 @@
-import identity from "lodash/identity";
-
 import { DeploymentParameters } from "../../../../types/deployer";
-import {
-  AccountRuntimeValue,
-  ArgumentType,
-  ContractFuture,
-  Future,
-  FutureType,
-  ModuleParameterRuntimeValue,
-  SolidityParameterType,
-} from "../../../../types/module";
-import { assertIgnitionInvariant } from "../../../utils/assertions";
-import { replaceWithinArg } from "../../../utils/replace-within-arg";
-import { resolveFromAddress } from "../../../utils/resolve-from-address";
-import { resolveModuleParameter } from "../../../utils/resolve-module-parameter";
+import { Future, FutureType } from "../../../../types/module";
 import { DeploymentState } from "../../types/deployment-state";
 import {
   DeploymentExecutionStateInitializeMessage,
   JournalMessage,
   JournalMessageType,
 } from "../../types/messages";
-import { findAddressForContractFuture } from "../../views/find-address-for-contract-future-by-id";
-import { findResultForFutureById } from "../../views/find-result-for-future-by-id";
+
+import {
+  resolveAddress,
+  resolveArgs,
+  resolveLibraries,
+  resolveValue,
+} from "./future-resolvers";
 
 export function buildInitializeMessageFor(
   future: Future,
@@ -41,15 +32,15 @@ export function buildInitializeMessageFor(
           dependencies: [...future.dependencies].map((f) => f.id),
           artifactFutureId: future.id,
           contractName: future.contractName,
-          constructorArgs: _resolveArgs(
+          constructorArgs: resolveArgs(
             future.constructorArgs,
             deploymentState,
             deploymentParameters,
             accounts
           ),
-          libraries: _resolveLibraries(future.libraries, deploymentState),
-          value: _resolveValue(future.value, deploymentParameters),
-          from: _resolveAddress(future.from, accounts),
+          libraries: resolveLibraries(future.libraries, deploymentState),
+          value: resolveValue(future.value, deploymentParameters),
+          from: resolveAddress(future.from, accounts),
         };
 
       return namedContractInitMessage;
@@ -94,70 +85,4 @@ export function buildInitializeMessageFor(
       throw new Error("Not implemented yet: FutureType.SEND_DATA case");
     }
   }
-}
-
-function _resolveValue(
-  givenValue: bigint | ModuleParameterRuntimeValue<bigint>,
-  deploymentParameters: DeploymentParameters
-): bigint {
-  if (typeof givenValue === "bigint") {
-    return givenValue;
-  }
-
-  const moduleParam = resolveModuleParameter(givenValue, {
-    deploymentParameters,
-  });
-
-  assertIgnitionInvariant(
-    typeof moduleParam === "bigint",
-    "Module parameter used as value must be a bigint"
-  );
-
-  return moduleParam;
-}
-
-function _resolveArgs(
-  args: ArgumentType[],
-  deploymentState: DeploymentState,
-  deploymentParameters: DeploymentParameters,
-  accounts: string[]
-): SolidityParameterType[] {
-  const replace = (arg: ArgumentType) =>
-    replaceWithinArg<SolidityParameterType>(arg, {
-      bigint: identity,
-      future: (f) => {
-        return findResultForFutureById(deploymentState, f.id);
-      },
-      accountRuntimeValue: (arv) => {
-        return accounts[arv.accountIndex];
-      },
-      moduleParameterRuntimeValue: (mprv) => {
-        return resolveModuleParameter(mprv, { deploymentParameters });
-      },
-    });
-
-  return args.map(replace);
-}
-
-function _resolveAddress(
-  from: string | AccountRuntimeValue | undefined,
-  accounts: string[]
-) {
-  if (from === undefined) {
-    return undefined;
-  }
-
-  return resolveFromAddress(from, { accounts });
-}
-
-function _resolveLibraries(
-  libraries: Record<string, ContractFuture<string>>,
-  deploymentState: DeploymentState
-) {
-  return Object.fromEntries(
-    Object.entries(libraries).map(([key, lib]) => [
-      key,
-      findAddressForContractFuture(deploymentState, lib.id),
-    ])
-  );
 }
